+++ src/simulation/entities/Fluid.hh
@@ -3,10 +3,10 @@
 namespace sim {
 
 template<typename T>
-Fluid<T>::Fluid(unsigned int id_, T density_, T viscosity_) : id(id_), density(density_), viscosity(viscosity_) { ++fluidCounter; }
+Fluid<T>::Fluid(size_t id_, T density_, T viscosity_) : id(id_), density(density_), viscosity(viscosity_) { ++fluidCounter; }
 
 template<typename T>
-Fluid<T>::Fluid(unsigned int id_, T density_, T viscosity_, std::string name_) : Fluid(id_, density_, viscosity_) {
+Fluid<T>::Fluid(size_t id_, T density_, T viscosity_, std::string name_) : Fluid(id_, density_, viscosity_) {
     this->name = std::move(name_);
 }
