+++ src/simulation/models/MixingModels.h
@@ -186,7 +186,7 @@ public:
      * @param[in] sim pointer to the simulation.
      * @param[in] mixtures reference to the unordered map of mixtures.
     */
-    virtual void updateMixtures(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<int, std::shared_ptr<Mixture<T>>>& mixtures) = 0;
+    virtual void updateMixtures(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& mixtures) = 0;
 
     virtual bool isInstantaneous() = 0;
 
@@ -222,7 +222,7 @@ public:
      * @param[in] sim pointer to the simulation.
      * @param[in] mixtures reference to the unordered map of mixtures.
     */
-    void updateMixtures(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<int, std::shared_ptr<Mixture<T>>>& mixtures) override;
+    void updateMixtures(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& mixtures) override;
 
     /**
      * @brief Move mixtures according to the timestep
@@ -238,7 +238,7 @@ public:
      * @param[in] network Pointer to the network of the simulation.
      * @param[in] mixtures Reference to collection containing all mixtures in the simulation
      */
-    void calculateMembraneExchange(T timeStep, AbstractMembrane<T>* sim, arch::Network<T>* network, const std::unordered_map<int, std::shared_ptr<Mixture<T>>>& mixtures);
+    void calculateMembraneExchange(T timeStep, AbstractMembrane<T>* sim, arch::Network<T>* network, const std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& mixtures);
 
     /**
      * @brief Calculate and store the mixtures flowing into all nodes
@@ -252,7 +252,7 @@ public:
      * @param[in] sim Pointer to the simulation.
      * @param[in] mixtures Unordered map of the mixtures in the system.
     */
-    void generateNodeOutflow(AbstractMixing<T>* sim, std::unordered_map<int, std::shared_ptr<Mixture<T>>>& mixtures);
+    void generateNodeOutflow(AbstractMixing<T>* sim, std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& mixtures);
 
     /**
      * @brief Add the node outflow as inflow to the channels
@@ -261,7 +261,7 @@ public:
      * @param[in] network Pointer to the network.
      * @param[in] mixtures Unordered map of the mixtures in the system.
     */
-    void updateChannelInflow(T timeStep, AbstractMixing<T>* sim, arch::Network<T>* network, std::unordered_map<int, std::shared_ptr<Mixture<T>>>& mixtures);
+    void updateChannelInflow(T timeStep, AbstractMixing<T>* sim, arch::Network<T>* network, std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& mixtures);
 
     /**
      * @brief Remove mixtures that have 'outflowed' their channel
@@ -326,7 +326,7 @@ public:
      * @param[in] sim pointer to the simulation.
      * @param[in] mixtures reference to the unordered map of mixtures.
     */
-    void updateMixtures(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<int, std::shared_ptr<Mixture<T>>>& mixtures) override;
+    void updateMixtures(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& mixtures) override;
 
     /**
      * @brief Propagate the mixtures and check if a mixtures reaches channel end.
@@ -336,7 +336,7 @@ public:
     /**
      * @brief Generate a new inflow in case a mixture has reached channel end. Invoked by updateNodeInflow.
     */
-    void generateInflows(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<int, std::shared_ptr<Mixture<T>>>& mixtures);
+    void generateInflows(T timeStep, arch::Network<T>* network, AbstractMixing<T>* sim, std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& mixtures);
 
     void topologyAnalysis(arch::Network<T>* network, int nodeId);
 
@@ -352,7 +352,7 @@ public:
     std::tuple<std::function<T(T)>,std::vector<T>, T> getAnalyticalSolutionFunction(T channelLength, T channelWidth, int resolution, T pecletNr, const std::vector<FlowSectionInput<T>>& parameters, std::function<T(T)> fConstant);
 
     std::tuple<std::function<T(T)>,std::vector<T>, T> getAnalyticalSolutionTotal(T channelLength, T currChannelFlowRate, T channelWidth, int resolution, int speciesId, T pecletNr, 
-        const std::vector<FlowSection<T>>& flowSections, std::unordered_map<int, std::shared_ptr<Mixture<T>>>& diffusiveMixtures);
+        const std::vector<FlowSection<T>>& flowSections, std::unordered_map<size_t, std::shared_ptr<Mixture<T>>>& diffusiveMixtures);
 
     void clean(arch::Network<T>* network);
